+++ litigant_portal/app/models/__init__.py
@@ -1,4 +1,5 @@
-from .admin import *  # noqa: F403
-from .assistant import *  # noqa: F403
 from .chat_engine import *  # noqa: F403
+from .site import *  # noqa: F403
+from .topic_flow import *  # noqa: F403
+from .upload import *  # noqa: F403
 from .user import *  # noqa: F403
