+++ litigant_portal/app/models/site.py
@@ -64,30 +64,3 @@ class SiteMembership(BaseModel):
                 fields=["user", "site"], name="unique_site_membership"
             )
         ]
-
-
-class Topic(BaseModel):
-    """A legal topic the app supports."""
-
-    id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
-    site = models.ForeignKey(
-        Site,
-        on_delete=models.CASCADE,
-        related_name="topics",
-    )
-    slug = models.SlugField(max_length=64)
-    title = models.CharField(max_length=255)
-    subtitle = models.CharField(max_length=255, blank=True)
-    description = models.CharField(max_length=255, blank=True)
-    icon = models.CharField(max_length=64, blank=True)
-    meta_description = models.TextField(blank=True)
-    prompts = models.JSONField(default=list, blank=True)
-    order = models.PositiveIntegerField(default=0)
-
-    class Meta:
-        ordering = ["order", "created_at"]
-        constraints = [
-            models.UniqueConstraint(
-                fields=["site", "slug"], name="unique_site_topic_slug"
-            )
-        ]
