+++ litigant_portal/app/models/topic_flow.py
@@ -1,0 +1,34 @@
+import uuid
+
+from django.conf import settings
+from django.db import models
+
+from .base import BaseModel
+from .choices import AI_MODEL_CHOICES, JurisdictionLevel, State
+from .site import Site
+
+class Topic(BaseModel):
+    """A legal topic the app supports."""
+
+    id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
+    site = models.ForeignKey(
+        Site,
+        on_delete=models.CASCADE,
+        related_name="topics",
+    )
+    slug = models.SlugField(max_length=64)
+    title = models.CharField(max_length=255)
+    subtitle = models.CharField(max_length=255, blank=True)
+    description = models.CharField(max_length=255, blank=True)
+    icon = models.CharField(max_length=64, blank=True)
+    meta_description = models.TextField(blank=True)
+    prompts = models.JSONField(default=list, blank=True)
+    order = models.PositiveIntegerField(default=0)
+
+    class Meta:
+        ordering = ["order", "created_at"]
+        constraints = [
+            models.UniqueConstraint(
+                fields=["site", "slug"], name="unique_site_topic_slug"
+            )
+        ]
